+++ packages/apollo-react-relay-duct-tape-compiler/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log - @graphitation/apollo-react-relay-duct-tape-compiler
 
-<!-- This log was last generated on Tue, 08 Apr 2025 15:39:00 GMT and should not be manually modified. -->
+<!-- This log was last generated on Thu, 29 May 2025 15:22:30 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 1.7.2
+
+Thu, 29 May 2025 15:22:30 GMT
+
+### Patches
+
+- Bump @graphitation/supermassive to v3.11.0
 
 ## 1.7.1